Some issues with editing Dinners
Saves on the local website, doesn't persist into the database. I had a
feeling this would be an issue, not sure how to fix atm - it probably is
a configuration issue since I changed the way the DB is setup. Maybe try
making a brand new nerddinner project and trying once more tomo